% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_functions.R
\name{AlignAA}
\alias{AlignAA}
\title{AlignAA}
\usage{
AlignAA(Shh_fasta)
}
\arguments{
\item{file}{a fasta file assigned to a vector in your environment}
}
\description{
Amino Acid multiple sequence alignment using the MUSCLE algorithm of fasta formatted data
}
\examples{
using the built-in Shh_fasta data of sonic hedgehog signal molecule orthologs
file <- "link to raw data on git hub"
pcms_AlignAA(file) # the output of this will be the sequences aligned by the MUSCLE algorithm
}
\keyword{AA}
\keyword{alignment}
\keyword{alignment,}
\keyword{fasta,}
\keyword{multiple}
\keyword{peptide,}
\keyword{sequence}
