% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_functions.R
\name{plot_Tree}
\alias{plot_Tree}
\title{plot_Tree}
\usage{
plot_Tree(tree)
}
\description{
This function uses phyloTools to print out a maximum likelihood using phylogenetic data
}
\examples{
plot_Tree <- tree
}
\keyword{likelihood}
\keyword{maximum}
\keyword{tree,}
