% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonichedgehog-package.R
\docType{package}
\name{sonichedgehog-package}
\alias{sonichedgehog}
\alias{sonichedgehog-package}
\title{sonichedgehog: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/slcornett/sonichedgehog}
  \item Report bugs at \url{https://github.com/slcornett/sonichedgehog/issues}
}

}
\keyword{internal}
