% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_functions.R
\name{treeAA}
\alias{treeAA}
\title{treeAA}
\usage{
treeAA(ptree)
}
\arguments{
\item{ptree, }{phylo formatted data}
}
\description{
This function plots a phlogenetic tree
}
\examples{
fas_msa <- "MsaMultipleAlignment" phyloAA(fas_msa) #the output is phylo tree data
}
\keyword{phylogenetric}
\keyword{tree}
