% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_functions.R
\name{phyloAA}
\alias{phyloAA}
\title{phyloAA}
\usage{
phyloAA(fas_msa)
}
\arguments{
\item{msa, }{formatted multiple alignment data}
}
\description{
This function conducts the multiple alignment and distance calculations for phylogenetic data
}
\examples{
fas_msa <- "MsaMultipleAlignment" phyloAA(fas_msa) #the output is phylo tree data
}
\keyword{distance}
\keyword{msa,}
