% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonichedgehog-package.R
\docType{package}
\name{sonichedgehog-package}
\alias{sonichedgehog}
\alias{sonichedgehog-package}
\title{sonichedgehog: Phylogenetic Comparative Methods}
\description{
This package embodies the quickness of Sonic the Hedgehog in providing a quick method to conduct phylogenetic compaative analyses. Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/slcornett/sonichedgehog}
  \item Report bugs at \url{https://github.com/slcornett/sonichedgehog/issues}
}

}
\keyword{internal}
