% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcm_functions.R
\name{treeAA}
\alias{treeAA}
\title{treeAA}
\usage{
treeAA(ptree)
}
\arguments{
\item{ptree, }{phylo formatted data}
}
\description{
This function plots a phlogenetic tree
}
\examples{
ptree <- "phylo" #prinnts out tree
}
\keyword{phylogenetric}
\keyword{tree}
